+++ src/main/java/seedu/address/model/task/ReadOnlyDeadline.java
@@ -6,17 +6,9 @@ import seedu.address.model.tag.UniqueTagList;
  * A read-only immutable interface for a Deadline in the TaskManager.
  * Implementations should guarantee: details are present and not null, field values are validated.
  */
-public interface ReadOnlyDeadline {
-
-	Name getName();
+public interface ReadOnlyDeadline extends ReadOnlyTask {
 	
 	TaskDate getEndDate();
-
-    /**
-     * The returned TagList is a deep copy of the internal TagList,
-     * changes on the returned list will not affect the task's internal tags.
-     */
-    UniqueTagList getTags();
 
     /**
      * Returns true if both have the same state. (interfaces cannot override .equals)
@@ -24,18 +16,9 @@ public interface ReadOnlyDeadline {
     default boolean isSameStateAs(ReadOnlyDeadline other) {
         return other == this // short circuit if same object
                 || (other != null // this is first to avoid NPE below
-                && other.getName().equals(this.getName()) // state checks here onwards
+                && other.getName().equals(this.getName())
+                && other.getEndDate().equals(this.getEndDate())// state checks here onwards
                 );
     }
 
-    /**
-     * Formats the person as text, showing all contact details.
-     */
-    default String getAsText() {
-        final StringBuilder builder = new StringBuilder();
-        builder.append(getName())
-                .append(" Tags: ");
-        getTags().forEach(builder::append);
-        return builder.toString();
-    }
 }
