+++ src/main/java/seedu/address/model/task/ReadOnlyEvent.java
@@ -6,19 +6,11 @@ import seedu.address.model.tag.UniqueTagList;
  * A read-only immutable interface for an Event in the TaskManager.
  * Implementations should guarantee: details are present and not null, field values are validated.
  */
-public interface ReadOnlyEvent {
-
-	Name getName();
+public interface ReadOnlyEvent extends ReadOnlyTask {
 	
 	TaskDate getStartDate();
 	
 	TaskDate getEndDate();
-
-    /**
-     * The returned TagList is a deep copy of the internal TagList,
-     * changes on the returned list will not affect the task's internal tags.
-     */
-    UniqueTagList getTags();
 
     /**
      * Returns true if both have the same state. (interfaces cannot override .equals)
@@ -26,18 +18,10 @@ public interface ReadOnlyEvent {
     default boolean isSameStateAs(ReadOnlyEvent other) {
         return other == this // short circuit if same object
                 || (other != null // this is first to avoid NPE below
-                && other.getName().equals(this.getName()) // state checks here onwards
+                && other.getName().equals(this.getName())
+                && other.getStartDate().equals(this.getStartDate())
+                && other.getEndDate().equals(this.getEndDate())// state checks here onwards
                 );
     }
 
-    /**
-     * Formats the person as text, showing all contact details.
-     */
-    default String getAsText() {
-        final StringBuilder builder = new StringBuilder();
-        builder.append(getName())
-                .append(" Tags: ");
-        getTags().forEach(builder::append);
-        return builder.toString();
-    }
 }
