+++ src/main/java/seedu/address/model/task/ReadOnlyFloatingTask.java
@@ -6,15 +6,7 @@ import seedu.address.model.tag.UniqueTagList;
  * A read-only immutable interface for a FloatingTask in the TaskManager.
  * Implementations should guarantee: details are present and not null, field values are validated.
  */
-public interface ReadOnlyFloatingTask {
-
-    Name getName();
-
-    /**
-     * The returned TagList is a deep copy of the internal TagList,
-     * changes on the returned list will not affect the person's internal tags.
-     */
-    UniqueTagList getTags();
+public interface ReadOnlyFloatingTask extends ReadOnlyTask {
 
     /**
      * Returns true if both have the same state. (interfaces cannot override .equals)
@@ -24,17 +16,6 @@ public interface ReadOnlyFloatingTask {
                 || (other != null // this is first to avoid NPE below
                 && other.getName().equals(this.getName()) // state checks here onwards
                 );
-    }
-
-    /**
-     * Formats the person as text, showing all contact details.
-     */
-    default String getAsText() {
-        final StringBuilder builder = new StringBuilder();
-        builder.append(getName())
-                .append(" Tags: ");
-        getTags().forEach(builder::append);
-        return builder.toString();
     }
 
 }
