+++ src/main/java/seedu/address/model/task/ReadOnlyTask.java
@@ -1,0 +1,25 @@
+package seedu.address.model.task;
+
+import seedu.address.model.tag.UniqueTagList;
+
+public interface ReadOnlyTask {
+
+	Name getName();
+	
+	/**
+     * The returned TagList is a deep copy of the internal TagList,
+     * changes on the returned list will not affect the person's internal tags.
+     */
+    UniqueTagList getTags();
+    
+    /**
+     * Formats the person as text, showing all contact details.
+     */
+    default String getAsText() {
+        final StringBuilder builder = new StringBuilder();
+        builder.append(getName())
+                .append(" Tags: ");
+        getTags().forEach(builder::append);
+        return builder.toString();
+    }
+}
